Drop productlist & search controllers in favor of view.php
