+++ controllers/front/view.php
@@ -17,134 +17,321 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-use PrestaShop\PrestaShop\Adapter\Image\ImageRetriever;
-use PrestaShop\PrestaShop\Adapter\Product\PriceFormatter;
-use PrestaShop\PrestaShop\Adapter\Product\ProductColorsRetriever;
-use PrestaShop\PrestaShop\Core\Product\ProductListingPresenter;
 
-class BlockWishlistViewModuleFrontController extends ModuleFrontController
+use PrestaShop\Module\BlockWishList\Search\WishListProductSearchProvider;
+use PrestaShop\PrestaShop\Core\Product\Search\ProductSearchQuery;
+use PrestaShop\PrestaShop\Core\Product\Search\SortOrder;
+use PrestaShop\PrestaShop\Core\Product\Search\SortOrderFactory;
+
+/**
+ * View the content of a personal wishlist
+ */
+class BlockWishlistViewModuleFrontController extends ProductListingFrontController
 {
+    /**
+     * @var BlockWishList
+     */
+    private $module;
+
+    /**
+     * @var WishList
+     */
+    protected $wishlist;
+
+    /**
+     * @var string
+     */
+    private $page_name;
+
     public function __construct()
     {
+        /** @var BlockWishList $module */
+        $module = Module::getInstanceByName('blockwishlist');
+        $this->module = $module;
+
+        if (empty($this->module->active)) {
+            Tools::redirect('index');
+        }
+
+        $this->page_name = 'module-' . $this->module->name . '-' . Dispatcher::getInstance()->getController();
+
         parent::__construct();
-        $this->context = Context::getContext();
+
+        $this->controller_type = 'modulefront';
     }
 
+    /**
+     * {@inheritdoc}
+     */
+    public function init()
+    {
+        $id_wishlist = $this->getWishlistId();
+        $this->wishlist = new WishList($id_wishlist);
+
+        if (false === Validate::isLoadedObject($this->wishlist)) {
+            Tools::redirect('index.php?controller=404');
+        }
+
+        parent::init();
+
+        if (false === $this->checkReadAccessToWishlist($this->wishlist)) {
+            header('HTTP/1.1 403 Forbidden');
+            header('Status: 403 Forbidden');
+            $this->errors[] = $this->trans(
+                'You do not have access to this wishlist.',
+                [],
+                'Modules.BlockWishList.Shop'
+            );
+            $this->setTemplate('errors/forbidden');
+            return;
+        }
+
+        $this->context->smarty->assign(
+            [
+                'id' => $id_wishlist,
+                'wishlistName' => $this->wishlist->name,
+                'isGuest' => !$this->checkWriteAccessToWishlist($this->wishlist),
+                'url' => Context::getContext()->link->getModuleLink('blockwishlist', 'view', $this->getAccessParams()),
+                'wishlistsLink' => Context::getContext()->link->getModuleLink('blockwishlist', 'lists'),
+                'deleteProductUrl' => Context::getContext()->link->getModuleLink('blockwishlist', 'action', ['action' => 'deleteProductFromWishlist']),
+            ]
+        );
+
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function initContent()
     {
         parent::initContent();
-        $token = Tools::getValue('token');
 
-        $module = new BlockWishList();
+        if (false === $this->checkReadAccessToWishlist($this->wishlist)) {
+            return;
+        }
 
-        if (true === empty($token)) {
-            $this->setTemplate('module:blockwishlist/views/templates/pages/view.tpl');
+        $this->context->controller->registerJavascript(
+            'blockwishlistController',
+            'modules/blockwishlist/public/productslist.bundle.js',
+            [
+              'priority' => 200,
+            ]
+        );
 
+        $this->doProductSearch(
+            '../../../modules/blockwishlist/views/templates/pages/products-list.tpl',
+            [
+                'entity' => 'wishlist_product',
+                'id_wishlist' => $this->wishlist->id,
+            ]
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getListingLabel()
+    {
+        return $this->trans(
+            'WishList: %wishlist_name%',
+            ['%wishlist_name%' => $this->wishlist->name],
+            'Modules.BlockWishList.Shop'
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getProductSearchQuery()
+    {
+        $query = new ProductSearchQuery();
+        $query->setSortOrder(
+            new SortOrder(
+                'product',
+                Tools::getProductsOrder('by'),
+                Tools::getProductsOrder('way')
+            )
+        );
+
+        return $query;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getDefaultProductSearchProvider()
+    {
+        return new WishListProductSearchProvider(
+            Db::getInstance(),
+            $this->wishlist,
+            new SortOrderFactory($this->getTranslator())
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getAjaxProductSearchVariables()
+    {
+        $data = parent::getAjaxProductSearchVariables();
+
+        $context = parent::getProductSearchContext();
+        $query = $this->getProductSearchQuery();
+        $provider = $this->getDefaultProductSearchProvider();
+
+        $resultsPerPage = (int) Tools::getValue('resultsPerPage');
+        if ($resultsPerPage <= 0) {
+            $resultsPerPage = Configuration::get('PS_PRODUCTS_PER_PAGE');
+        }
+
+        // we need to set a few parameters from back-end preferences
+        $query
+            ->setResultsPerPage($resultsPerPage)
+            ->setPage(max((int) Tools::getValue('page'), 1))
+        ;
+
+        // set the sort order if provided in the URL
+        if (($encodedSortOrder = Tools::getValue('order'))) {
+            $query->setSortOrder(SortOrder::newFromString(
+                $encodedSortOrder
+            ));
+        }
+
+        // get the parameters containing the encoded facets from the URL
+        $encodedFacets = Tools::getValue('q');
+
+        $query->setEncodedFacets($encodedFacets);
+
+        $result = $provider->runQuery(
+            $context,
+            $query
+        );
+
+        if (!$result->getCurrentSortOrder()) {
+            $result->setCurrentSortOrder($query->getSortOrder());
+        }
+
+        // prepare the products
+        $products = $this->prepareMultipleProductsForTemplate(
+            $result->getProducts()
+        );
+
+        // render the facets
+        // with the core
+        $rendered_facets = $this->renderFacets(
+            $result
+        );
+        $rendered_active_filters = $this->renderActiveFilters(
+            $result
+        );
+
+        $pagination = $this->getTemplateVarPagination(
+            $query,
+            $result
+        );
+
+        // prepare the sort orders
+        // note that, again, the product controller is sort-orders
+        // agnostic
+        // a module can easily add specific sort orders that it needs
+        // to support (e.g. sort by "energy efficiency")
+        $sort_orders = $this->getTemplateVarSortOrders(
+            $result->getAvailableSortOrders(),
+            $query->getSortOrder()->toString()
+        );
+
+        $sort_selected = false;
+        if (!empty($sort_orders)) {
+            foreach ($sort_orders as $order) {
+                if (isset($order['current']) && true === $order['current']) {
+                    $sort_selected = $order['label'];
+
+                    break;
+                }
+            }
+        }
+
+        $searchVariables = [
+            'result' => $result,
+            'label' => $this->getListingLabel(),
+            'products' => $products,
+            'sort_orders' => $sort_orders,
+            'sort_selected' => $sort_selected,
+            'pagination' => $pagination,
+            'rendered_facets' => $rendered_facets,
+            'rendered_active_filters' => $rendered_active_filters,
+            'js_enabled' => $this->ajax,
+            'current_url' => $this->updateQueryString([
+                'q' => $result->getEncodedFacets(),
+            ]),
+        ];
+
+        Hook::exec('filterProductSearch', ['searchVariables' => &$searchVariables]);
+        Hook::exec('actionProductSearchAfter', $searchVariables);
+
+        return $searchVariables;
+    }
+
+    /**
+     * Depending on the parameters sent, checks if the current visitor may reach the page
+     * 
+     * @return int|false
+     */
+    private function getWishlistId()
+    {
+        if (Tools::getIsset('id_wishlist')) {
+            return (int) Tools::getValue('id_wishlist');
+        }
+
+        if (Tools::getIsset('token')) {
+            $wishlistData = WishList::getByToken(
+                Tools::getValue('token')
+            );
+
+            if (!empty($wishlistData['id_wishlist'])) {
+                return $wishlistData['id_wishlist'];
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @return bool
+     */
+    private function checkReadAccessToWishlist(WishList $wishlist)
+    {
+        // Wishlist is shared
+        if (!empty($wishlist->token) && Tools::getIsset('token')) {
+            return true;
+        }
+
+        return $this->checkWriteAccessToWishlist($wishlist);
+    }
+
+        /**
+     * @return bool
+     */
+    private function checkWriteAccessToWishlist(WishList $wishlist)
+    {
+        if (false === Validate::isLoadedObject($this->context->customer)) {
             return false;
         }
 
-        $wishlist = WishList::getByToken($token);
-        WishList::refreshWishList($wishlist['id_wishlist']);
+        return ((int) $wishlist->id_customer) === $this->context->customer->id;
+    }
 
-        $products = WishList::getProductByIdCustomer(
-            (int) $wishlist['id_wishlist'],
-            (int) $wishlist['id_customer'],
-            $this->context->language->id,
-            null,
-            true
-        );
-
-        $nb_products = count($products);
-        $priority_names = [0 => $module->l('High'), 1 => $module->l('Medium'), 2 => $module->l('Low')];
-
-        for ($i = 0; $i < $nb_products; ++$i) {
-            $product = new Product((int) $products[$i]['id_product'], true, $this->context->language->id);
-
-            $products[$i]['priority_name'] = $priority_names[$products[$i]['priority']];
-            $quantity = Product::getQuantity((int) $products[$i]['id_product'], $products[$i]['id_product_attribute']);
-            $products[$i]['attribute_quantity'] = $quantity;
-            $products[$i]['product_quantity'] = $quantity;
-            $products[$i]['allow_oosp'] = $product->isAvailableWhenOutOfStock((int) $product->out_of_stock);
-
-            if ($products[$i]['id_product_attribute'] != 0) {
-                $combination_imgs = $product->getCombinationImages($this->context->language->id);
-                if (isset($combination_imgs[$products[$i]['id_product_attribute']][0])) {
-                    $products[$i]['cover'] = $product->id . '-' . $combination_imgs[$products[$i]['id_product_attribute']][0]['id_image'];
-                } else {
-                    $cover = Product::getCover($product->id);
-                    $products[$i]['cover'] = $product->id . '-' . $cover['id_image'];
-                }
-            } else {
-                $images = $product->getImages($this->context->language->id);
-                foreach ($images as $image) {
-                    if ($image['cover']) {
-                        $products[$i]['cover'] = $product->id . '-' . $image['id_image'];
-                        break;
-                    }
-                }
-            }
-            if (!isset($products[$i]['cover'])) {
-                $products[$i]['cover'] = $this->context->language->iso_code . '-default';
-            }
-
-            $products[$i]['bought'] = false;
+    /**
+     * @return array
+     */
+    private function getAccessParams()
+    {
+        if (Tools::getIsset('token')) {
+            return ['token' => Tools::getValue('token')];
         }
-
-        $assembler = new ProductAssembler($this->context);
-        $presenterFactory = new ProductPresenterFactory($this->context);
-        $presentationSettings = $presenterFactory->getPresentationSettings();
-        $presenter = new ProductListingPresenter(
-            new ImageRetriever(
-                $this->context->link
-            ),
-            $this->context->link,
-            new PriceFormatter(),
-            new ProductColorsRetriever(),
-            $this->context->getTranslator()
-        );
-        $products_for_template = [];
-
-        if (is_array($products)) {
-            foreach ($products as $rawProduct) {
-                $products_for_template[] = $presenter->present(
-                    $presentationSettings,
-                    $assembler->assembleProduct($rawProduct),
-                    $this->context->language
-                );
-            }
+        if (Tools::getIsset('id_wishlist')) {
+            return ['id_wishlist' => Tools::getValue('id_wishlist')];
         }
-
-        WishList::incCounter((int) $wishlist['id_wishlist']);
-        $ajax = Configuration::get('PS_BLOCK_CART_AJAX');
-
-        $wishlists = WishList::getByIdCustomer((int) $wishlist['id_customer']);
-
-        foreach ($wishlists as $key => $item) {
-            if ($item['id_wishlist'] == $wishlist['id_wishlist']) {
-                unset($wishlists[$key]);
-                break;
-            }
-
-            $this->context->controller->registerJavascript(
-                'blockwishlistController',
-                'modules/blockwishlist/public/productslist.bundle.js',
-                [
-                  'priority' => 200,
-                ]
-            );
-
-            $this->context->smarty->assign(
-                [
-                    'current_wishlist' => $wishlist,
-                    'token' => $token,
-                    'ajax' => (($ajax == 1) ? '1' : '0'),
-                    'wishlists' => $wishlists,
-                    'isShared' => true,
-                    'products' => $products_for_template,
-                ]
-            );
-
-            $this->setTemplate('module:blockwishlist/views/templates/pages/view.tpl');
-        }
+        return [];
     }
 }
